+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 5.2.1
 
 - Fixes a bug, when adding invalid modules to the app
+- Fixes a bug, where not awaiting set oblique collection lead to unexpected state
 
 ### 5.2.0